Move db scripts into their own folder.
