+++ src/scripts/insights/index.mjs
@@ -1,4 +1,4 @@
-import { getAllFromObjectStore, getAllFromCloud } from './db.mjs';
+import { getAllFromObjectStore, getAllFromCloud } from './db/index.mjs';
 import { updateBackLink, formatCurrency } from './helpers/index.mjs';
 
 updateBackLink();
